+++ src/common/hooks/useOrder/useOrder.js
@@ -1,38 +1,42 @@
 import { useQuery } from 'react-query';
 
 import { useOkapiKy } from '@folio/stripes/core';
-
 import { ORDERS_API } from '@folio/stripes-acq-components';
 
 // tries to fetch order by get, if error comes from back-end - fetch from collection api
-export const useOrder = (orderId) => {
+export const useOrder = (orderId, options = {}) => {
+  const {
+    enabled = true,
+    fiscalYearId,
+    ...queryOptions
+  } = options;
+
   const ky = useOkapiKy();
 
-  const searchParams = {
-    query: `id==${orderId}`,
-  };
-
-  const {
-    data,
-    isLoading,
-    refetch,
-  } = useQuery(
-    ['ui-orders', 'order', orderId],
-    async ({ signal }) => {
+  const { data, ...rest } = useQuery({
+    queryKey: ['ui-orders', 'order', orderId, fiscalYearId],
+    queryFn: async ({ signal }) => {
       try {
-        return ky.get(`${ORDERS_API}/${orderId}`, { signal }).json();
+        const searchParams = fiscalYearId ? { fiscalYearId } : undefined;
+
+        return ky.get(`${ORDERS_API}/${orderId}`, { searchParams, signal }).json();
       } catch {
-        const { purchaseOrders } = await ky.get(`${ORDERS_API}`, { searchParams, signal }).json();
+        const searchParams = {
+          query: `id==${orderId}`,
+          ...(fiscalYearId ? { fiscalYearId } : {}),
+        };
+
+        const { purchaseOrders } = await ky.get(ORDERS_API, { searchParams, signal }).json();
 
         return purchaseOrders[0] || {};
       }
     },
-    { enabled: Boolean(orderId) },
-  );
+    enabled: enabled && Boolean(orderId),
+    ...queryOptions,
+  });
 
   return ({
     order: data,
-    isLoading,
-    refetch,
+    ...rest,
   });
 };