+++ src/common/hooks/useOrderLines/useOrderLines.js
@@ -1,0 +1,50 @@
+import { useQuery } from 'react-query';
+
+import {
+  useNamespace,
+  useOkapiKy,
+} from '@folio/stripes/core';
+import {
+  CQLBuilder,
+  LIMIT_MAX,
+  LINES_API,
+} from '@folio/stripes-acq-components';
+
+const DEFAULT_DATA = [];
+
+export const useOrderLines = (options = {}) => {
+  const {
+    enabled = true,
+    searchParams: params = {},
+    tenantId,
+    ...queryOptions
+  } = options;
+
+  const ky = useOkapiKy({ tenant: tenantId });
+  const [namespace] = useNamespace({ key: 'order-lines' });
+
+  const { data, ...rest } = useQuery({
+    queryKey: [namespace, tenantId, ...Object.values(params)],
+    queryFn: async ({ signal }) => {
+      const searchParams = {
+        limit: LIMIT_MAX,
+        query: (
+          new CQLBuilder()
+            .allRecords()
+            .build()
+        ),
+        ...params,
+      };
+
+      return ky.get(LINES_API, { searchParams, signal }).json();
+    },
+    enabled,
+    ...queryOptions,
+  });
+
+  return ({
+    orderLines: data?.poLines || DEFAULT_DATA,
+    totalRecords: data?.totalRecords,
+    ...rest,
+  });
+};