+++ src/components/PurchaseOrder/POInvoices/POInvoicesContainer.js
@@ -1,31 +1,38 @@
 import PropTypes from 'prop-types';
 
-import { Accordion, Loading } from '@folio/stripes/components';
+import {
+  Accordion,
+  Loading,
+} from '@folio/stripes/components';
 
 import { ACCORDION_ID } from '../../POLine/const';
 import POInvoices from './POInvoices';
 import { useRelatedInvoices } from './useRelatedInvoices';
 
-const POInvoicesContainer = ({ label, orderInvoicesIds }) => {
-  const { isLoading, orderInvoices } = useRelatedInvoices(orderInvoicesIds);
+const POInvoicesContainer = ({
+  isLoading,
+  label,
+  orderInvoicesIds,
+}) => {
+  const {
+    isLoading: isInvoicesLoading,
+    orderInvoices,
+  } = useRelatedInvoices(orderInvoicesIds);
 
   return (
     <Accordion
       label={label}
       id={ACCORDION_ID.relatedInvoices}
     >
-      {isLoading ? <Loading /> : <POInvoices orderInvoices={orderInvoices} />}
+      {(isLoading || isInvoicesLoading) ? <Loading /> : <POInvoices orderInvoices={orderInvoices} />}
     </Accordion>
   );
 };
 
 POInvoicesContainer.propTypes = {
+  isLoading: PropTypes.bool,
+  label: PropTypes.node.isRequired,
   orderInvoicesIds: PropTypes.arrayOf(PropTypes.string),
-  label: PropTypes.object.isRequired,
-};
-
-POInvoicesContainer.defaultProps = {
-  orderInvoicesIds: [],
 };
 
 export default POInvoicesContainer;